% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combined_tourn.r
\name{active.passive.matchings}
\alias{active.passive.matchings}
\title{Matching with active a and passive strategies}
\usage{
active.passive.matchings(active.strat, passive.strat, game)
}
\arguments{
\item{active.strat}{named list of active strategies. Each strategy is a function}

\item{passive.strat}{named list of passive strategies}

\item{game}{game object}
}
\description{
Each active strat plays agains each active strat and each passive strat.
Passive strats do not play against each other.
}
\details{
Can be useful if we already have tournament data in which the passive strats
play against each other and we want it to add the active strats to this
existing tournament data.
}

