% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{coop.high.C}
\alias{coop.high.C}
\title{Cooperates if cooperation makes sense in a naive way}
\usage{
coop.high.C(obs, i, t, ...)
}
\description{
A strategy for the iterated prisoners dilemma
It chooses the action with the highest sum value of the payoff matrix (i.e. always.defect in the case of the standard version);
see make.pd.game
}
