% Generated by roxygen2 (4.0.1): do not edit by hand
\name{debug.restore}
\alias{debug.restore}
\title{restore objects if and only if called from the global environment}
\usage{
debug.restore(strat, i, t, stage = NULL)
}
\description{
restore objects if and only if called from the global environment
}

