% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle_of_strategies.r
\name{debug.store}
\alias{debug.store}
\title{store objects when called from a function}
\usage{
debug.store(strat, i, t, stage = NULL, do.store = TRUE)
}
\description{
store objects when called from a function
}
