% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strattourn_utils.r
\name{distance.to.previous.event}
\alias{distance.to.previous.event}
\title{x is a vector of 0s and 1s (or FALSE and TRUE)}
\usage{
distance.to.previous.event(x)
}
\description{
the function returns a vector of length(x) containing for each index i the distance to the previous entry of a 1, if x[i]=1 we return 0. If the vector starts with 0's, we set the distance for these 0 entries to Inf.
}

