% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargaining_game.r
\name{examples.bargaining.game}
\alias{examples.bargaining.game}
\title{The code inside can be used to explore the behavior of strategies for the PD game}
\usage{
examples.bargaining.game()
}
\description{
The code inside can be used to explore the behavior of strategies for the PD game
}
