% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{examples.pd}
\alias{examples.pd}
\title{The code inside can be used to explore the behavior of strategies for the PD game}
\usage{
examples.pd()
}
\description{
The code inside can be used to explore the behavior of strategies for the PD game
}
