% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_agreement_game.r
\name{give.generously.1}
\alias{give.generously.1}
\title{Example strategy: give.generously.1 (more=1)}
\usage{
give.generously.1(obs, i, t, more = 1, ...)
}
\arguments{
\item{obs}{Observations of current know how and share of previous round as specified in the function make.bargaining.game}

\item{i}{Number of current player}

\item{t}{Current period}

\item{more}{How much know how should be given in addition to the shared know how of the other firm in the last period?}
}
\description{
Strategy which shares (in the research agreement game) as much as the other firm + more, if possible; in the first period everything
}
