% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotelling_game.r
\name{hotelling.profits}
\alias{hotelling.profits}
\title{Calculates the profits of a hotelling game}
\usage{
hotelling.profits(lower.bound = 0, upper.bound = 1, s = 1, t = 1,
  print.details = FALSE, choice1, choice2)
}
\arguments{
\item{lower.bound}{Customers are uniformly distributed in the intervall [lower.bound, upper.bound]}

\item{upper.bound}{Customers are uniformly distributed in the intervall [lower.bound, upper.bound]}

\item{s}{Prohibitive price of customer}

\item{t}{Factor determining how much the customer loathes distance}

\item{print.details}{When TRUE the function prints various details and intermediate calculations.}

\item{choice1}{choice of first company. The choice has to have the structure list(p=<numerical price>, l=<numerical location>).}

\item{choice2}{choice of first company. The choice has to have the structure list(p=<numerical price>, l=<numerical location>). Note that in contrast to some standard notations of the hotelling game the location choice of firm2 is given in absolute values and neither starting from the lower bound nor from the upper bound.}
}
\description{
We consider a hotelling game where two firms set price and location at the same time. The Utility Function of the uniformly distributed customers is 
U = s - t.distance*|x-l|-p.
}
