% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargaining_game.r
\name{human.player.bg}
\alias{human.player.bg}
\title{Waits for Input of Human Player in Bargaining Game}
\usage{
human.player.bg(obs, i, t, ...)
}
\description{
Waits for Input of Human Player in Bargaining Game
}
