% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotelling_game.r
\name{human.player.ht}
\alias{human.player.ht}
\title{Waits for Input of Human Player in Hotelling Game}
\usage{
human.player.ht(obs, i, t, ...)
}
\description{
Waits for Input of Human Player in Hotelling Game
}
