% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{human.player.pd}
\alias{human.player.pd}
\title{Waits for Input of Human Player in Prisoners Dilemma Game}
\usage{
human.player.pd(obs, i, t, ...)
}
\description{
Waits for Input of Human Player in Prisoners Dilemma Game
}
