% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_agreement_game.r
\name{human.player.ra}
\alias{human.player.ra}
\title{Waits for Input of Human Player in Research Agreement Game}
\usage{
human.player.ra(obs, i, t, ...)
}
\description{
Waits for Input of Human Player in Research Agreement Game
}
