% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rps_game.r
\name{human.player.rps}
\alias{human.player.rps}
\title{Waits for Input of Human Player in Rock-Papers-Scissors}
\usage{
human.player.rps(obs, i, t, ...)
}
\description{
Waits for Input of Human Player in Rock-Papers-Scissors
}
