\name{import.stage2.strats}
\alias{import.stage2.strats}
\title{Import Stage 1 strategies and Stage 2 answers}
\usage{
  import.stage2.strats(stage1.dir, stage2.dir, num.scen)
}
\arguments{
  \item{stage1.dir}{the directory in which the Stage 1 Rmd
  files are stored}

  \item{stage2.dir}{the directory in which the Stage 2 Rmd
  files are stored}

  \item{num.scen}{the number of scenarios}
}
\value{
  a list that contains a list for each scenario. The list
  for each scenario has a list strat that contains the
  strategies of all teams.
}
\description{
  Import Stage 1 strategies and Stage 2 answers
}

