% Generated by roxygen2 (4.0.1): do not edit by hand
\name{init.random.state}
\alias{init.random.state}
\title{Sets the random number state to a state specified by name}
\usage{
init.random.state(name = ".GLOBAL", seed = NULL)
}
\description{
Sets the random number state to a state specified by name
}

