% Generated by roxygen2 (4.0.1): do not edit by hand
\name{init.tournament}
\alias{init.tournament}
\title{Inits a tournament object}
\usage{
init.tournament(strat, game, matchings = NULL, score.fun = "u",
  team = NULL, rs.file = NULL, dir = getwd(), tourn.id = NULL,
  id.add = NULL)
}
\description{
Inits a tournament object
}

