% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.game.maker}
\alias{load.game.maker}
\title{Loads a game maker
A game maker is a function that can generate a game}
\usage{
\method{load}{game.maker}(file)
}
\description{
Loads a game maker
A game maker is a function that can generate a game
}

