% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{load.tournament}
\alias{load.tournament}
\title{Loads a tournament from a file}
\usage{
\method{load}{tournament}(file = paste0(tourn$tourn.id, ".tou"),
  tourn = NULL, path = getwd())
}
\description{
Loads a tournament from a file
}
