% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotelling_game.r
\name{make.hotelling.game}
\alias{make.hotelling.game}
\title{Generate a Hotelling Game}
\usage{
make.hotelling.game(lower.bound = 0, upper.bound = 1, s = 1,
  t.distance = 1, fix.location = FALSE, delta = 0.9, ...)
}
\arguments{
\item{lower.bound}{Customers are uniformly distributed in the intervall [lower.bound, upper.bound]}

\item{upper.bound}{Customers are uniformly distributed in the intervall [lower.bound, upper.bound]}

\item{s}{Prohibitive price of customer}

\item{t.distance}{Factor determining how much the customer loathes distance}

\item{fix.location}{If TRUE, then the location can only be set once}

\item{delta}{Probability of playing another round}
}
\description{
We consider two firms playing a repeated hotelling game, setting price and location. The Utility Function of the uniformly distributed customers is 
U = s - t.distance*|x-l|-p.
}

