\name{make.payment.message.pd.game}
\alias{make.payment.message.pd.game}
\title{Generate a (noisy) Prisoners' Dilemma game in which players can conduct messages and payments}
\usage{
  make.payment.message.pd.game(uCC = 1, uCD = -1, uDC = 2,
    uDD = 0, err.D.prob = 0, err.C.prob = 0,
    private.signals = FALSE)
}
\description{
  Generate a (noisy) Prisoners' Dilemma game in which
  players can conduct messages and payments
}

