\name{make.payment.message.sabotage.pd.game}
\alias{make.payment.message.sabotage.pd.game}
\title{Generate a (noisy) Prisoners' Dilemma game in which players can conduct messages and payments, as well as perform costly sabotage that makes the other player look like a defector}
\usage{
  make.payment.message.sabotage.pd.game(uCC = 1, uCD = -1,
    uDC = 2, uDD = 0, err.D.prob = 0, err.C.prob = 0,
    sabotage.cost = 0.3)
}
\description{
  Generate a (noisy) Prisoners' Dilemma game in which
  players can conduct messages and payments, as well as
  perform costly sabotage that makes the other player look
  like a defector
}

