% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.pd.game}
\alias{make.pd.game}
\title{Generate a (noisy) Prisoners' Dilemma game}
\usage{
make.pd.game(uCC = 1, uCD = -1, uDC = 2, uDD = 0, err.D.prob = 0,
  err.C.prob = 0, private.signals = FALSE, delta = 0.9, ...)
}
\description{
Generate a (noisy) Prisoners' Dilemma game
}

