% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.pg.game}
\alias{make.pg.game}
\title{Generate a public goods game}
\usage{
make.pg.game(n = 5, MPCR = 2/n, pun.cost = 1/3, max.x = 100,
  max.pun = 200, delta = 0.9, private.signals = FALSE, ...)
}
\description{
Generate a public goods game
}

