% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pg_cost_game.r
\name{make.pg_cost.game}
\alias{make.pg_cost.game}
\title{Generate a (noisy) Prisoners' Dilemma game}
\usage{
make.pg_cost.game(cost.min = 0.3, cost.max = 1.3, new.cost.prob = 0.5,
  digits = 0, delta = 0.95, ...)
}
\description{
Generate a (noisy) Prisoners' Dilemma game
}

