% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/research_agreement_game.r
\name{make.research.agreement.game}
\alias{make.research.agreement.game}
\title{Generate a research agreement game}
\usage{
make.research.agreement.game(lower.bound = 1, upper.bound = 5,
  multiplier = 1.5, delta = 0.9, ...)
}
\arguments{
\item{lower.bound}{At least this much know how is generated from each firm in each period}

\item{upper.bound}{At most this much know how is generated from each firm in each period}

\item{multiplier}{Non shared know how is multiplied by this value}

\item{delta}{Probability of playing another round}
}
\description{
We consider two firms which generate a random amount of know how each period. Each firm knows how much know how has been generated by itself, but does not know how much know how has been generated by the other firm. In each period they can decide to share a part of their know how with the other firm.
}

