% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rps_game.r
\name{make.rps.game}
\alias{make.rps.game}
\title{Generate a rock papers scissors game}
\usage{
make.rps.game(delta = NULL, T = NULL, cost.r = 0.75, cost.p = 0,
  cost.s = 0, ...)
}
\description{
Generate a rock papers scissors game
}
