% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/research_agreement_game.r
\name{most.of.it}
\alias{most.of.it}
\title{Example strategy: most.of.it}
\usage{
most.of.it(obs, i, t, ...)
}
\arguments{
\item{obs}{Observations of current know how and share of previous round as specified in the function make.bargaining.game}

\item{i}{Number of current player}

\item{t}{Current period}
}
\description{
An example strategy that always shares n-1 of its know-how in the research agreement game
}

