% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_solutions.r
\name{parse.strats.from.rmd}
\alias{parse.strats.from.rmd}
\title{Parse strategies from a rmd solution file}
\usage{
\method{parse}{strats.from.rmd}(file, num.scen)
}
\arguments{
\item{file}{the name of your rmd solution file}

\item{num.scen}{the number of scenarios}
}
\value{
A list with codes and information about your strategy. If the strategies cannot be correctly parsed, the function throws an error.
}
\description{
Parse strategies from a rmd solution file
}

