% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{random.action}
\alias{random.action}
\title{A strategy that randomly chooses an action}
\usage{
random.action(obs, i, t, ...)
}
\description{
A strategy for the iterated prisoners dilemma;
see make.pd.game
}
