% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.rep.game}
\alias{run.rep.game}
\title{Runs a repeated game}
\usage{
run.rep.game(delta = game$delta, game, strat, T.max = NULL,
  detailed.return = TRUE, strat.seed = NULL, game.seed = NULL,
  do.store = TRUE, strat.par = NULL, sample.delta = 1 - (1 - delta)/2,
  match.id = sample.int(2147483647, 1))
}
\arguments{
\item{delta}{discount factor}

\item{game}{the game object}

\item{strat}{a list of strategies}

\item{T.max}{optionally a maximum number of rounds}
}
\description{
Runs a repeated game
}

