% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.tournament}
\alias{run.tournament}
\title{Runs a tournament with R repetitions of each matching and add these rounds to the tournament objects}
\usage{
run.tournament(tourn, strat = tourn$strat, matchings = tourn$matchings,
  game = tourn$game, delta = game$delta, R = 5, LAPPLY = lapply,
  verbose = interactive() * 1, do.store = FALSE,
  matchings.fun = random.group.matchings,
  fixed.matchings = !is.null(matchings), weights = NULL, ...)
}
\description{
By setting backup.each.R to a number, say 10, a backup of the tournament will be created after each 10 repetitions
}

