% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{save.tournament}
\alias{save.tournament}
\title{Saves a tournament to a file}
\usage{
\method{save}{tournament}(tourn, path = getwd(), file = NULL,
  add.stats = FALSE)
}
\description{
Saves a tournament to a file
}
