% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show.tournament}
\alias{show.tournament}
\title{Analyse a tournament interactively in web browser}
\usage{
show.tournament(tourn = NULL, tourn.file = NULL, launch.browser = TRUE,
  file.path = getwd(), strat.shares = NULL)
}
\description{
Analyse a tournament interactively in web browser
}

