% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny_against_given.r
\name{showAgainstGiven}
\alias{showAgainstGiven}
\title{Analyse a tournament interactively in web browser}
\usage{
showAgainstGiven(tourn = NULL, tourn.file = NULL, launch.browser = TRUE,
  file.path = getwd(), strat.shares = NULL)
}
\description{
Analyse a tournament interactively in web browser
}

