% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{strange.defector}
\alias{strange.defector}
\title{Strategy from the tutorial without much meaning}
\usage{
strange.defector(obs, i, t, still.defect = 0, ...)
}
\description{
A strategy for the iterated prisoners dilemma;
see make.pd.game
}
