% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{tit.for.tat}
\alias{tit.for.tat}
\title{The famous tit.for.tat strategy: winner of Axelrod's original tournament}
\usage{
tit.for.tat(obs, i, t, ...)
}
\description{
A strategy for the iterated prisoners dilemma;
see make.pd.game
}
