% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotelling_game.r
\name{traveling.salesman}
\alias{traveling.salesman}
\title{Example strategy (Hotelling Game): traveling.salesman}
\usage{
traveling.salesman(obs, i, t, price = 0.5, speed = 0.05, ...)
}
\arguments{
\item{obs}{Observations of position and price of both players in the last round as specified in the function hotelling.profits}

\item{i}{Number of current player}

\item{t}{Current period}

\item{price}{Fixed price of the strategy}

\item{speed}{The strategy starts at 0 and moves each period by speed, afterwards it starts at the beginning.}
}
\description{
A strategy which "walks" through over the "beach" with a fixed price
}

