% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uglify.R
\name{uglify.strat}
\alias{uglify.strat}
\title{Makes a strategy (function) hard to read}
\usage{
uglify.strat(strat, keep.vars = names(formals(strat)), keep.funs = NULL,
  ...)
}
\description{
Purpose: This way we can give strategies to students
who can test their strategies against them. While in principle
one can write some code to make the functions easier readable
again, this is not straightforward. The effort seems fairly 
large. If students do that
}
