% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_code.r
\name{whitelist.check.strat}
\alias{whitelist.check.strat}
\title{Checks whether a given strategy passes a whitelist}
\usage{
whitelist.check.strat(strat)
}
\description{
Checks whether a given strategy passes a whitelist
}

