% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battle_of_strategies.r
\name{debug.restore}
\alias{debug.restore}
\title{restore objects if and only if called from the global environment}
\usage{
debug.restore(strat, i, t, stage = NULL)
}
\description{
restore objects if and only if called from the global environment
}

