% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_random_seed.r
\name{init.random.state}
\alias{init.random.state}
\title{Sets the random number state to a state specified by name}
\usage{
init.random.state(name = ".GLOBAL", seed = NULL)
}
\description{
Sets the random number state to a state specified by name
}

