% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{init.tournament}
\alias{init.tournament}
\title{Inits a tournament object}
\usage{
init.tournament(strat, game, matchings = NULL, score.fun = "u",
  team = NULL, rs.file = NULL, dir = getwd(), tourn.id = NULL,
  id.add = NULL, separate.round.data = TRUE, game.seeds = NULL)
}
\description{
Inits a tournament object
}
