% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle_of_strategies.r
\name{load.game.maker}
\alias{load.game.maker}
\title{Loads a game maker
A game maker is a function that can generate a game}
\usage{
\method{load}{game.maker}(file)
}
\description{
Loads a game maker
A game maker is a function that can generate a game
}
