% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgame.r
\name{make.pd.game}
\alias{make.pd.game}
\title{Generate an iterated Prisoners Dilemma game}
\usage{
make.pd.game(uCC = 1, uCD = -1, uDC = 2, uDD = 0, digits = 2,
  err.D.prob = 0, err.C.prob = 0, private.signals = FALSE, delta = 0.9,
  ...)
}
\arguments{
\item{uCC}{Utility in the case 'both players cooperate'. If specified in form of a vector (e.g. c(0,2)), then a random value is uniformly drawn each round and presented to the strategies.}

\item{uCD}{Utility in the case 'I cooperate, but the other one doesn't'. If specified in form of a vector (e.g. c(0,2)), then a random value is uniformly drawn each round and presented to the strategies.}

\item{uDC}{Utility in the case 'I defect, but the other one cooperates'. If specified in form of a vector (e.g. c(0,2)), then a random value is uniformly drawn each round and presented to the strategies.}

\item{uDD}{Utility in the case 'both players defect'. If specified in form of a vector (e.g. c(0,2)), then a random value is uniformly drawn each round and presented to the strategies.}

\item{digits}{If the costs are drawn from a uniform distribution: To how many digits should the draw be rounded?}

\item{err.D.prob}{Probability that a cooperation is seen as a defection, i.e. the propability to erroneous see a D}

\item{err.C.prob}{Probability that a defection is seen as a cooperation, i.e. the propability to erroneous see a C}

\item{private.signals}{If TRUE, then every player always sees his own action as originaly intended. He can no longer see whether his action in the previous round was due to an error.}

\item{delta}{Probability of playing another round}
}
\description{
Generate an iterated Prisoners Dilemma game
}
