% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pos.in.seq.of.ones}
\alias{pos.in.seq.of.ones}
\title{x is a vector of 0s and 1s (or FALSE and TRUE)}
\usage{
pos.in.seq.of.ones(x)
}
\description{
the function returns a vector of length(x) containing for each 1 the position within each sequence of 1s and a 0 for each 0.
}

